+++ src/JlWrap/any.jl
@@ -389,6 +389,49 @@ end
 
 pyjlany_hash(self) = pyint(hash(self))
 
+const PYNUMPYDTYPE = IdDict{Type,Py}()
+
+function pyjlany_numpy_dtype(self::Type)
+    ans = get!(PYNUMPYDTYPE, self) do
+        typestr, descr = pytypestrdescr(self)
+        # unsupported type
+        if typestr == ""
+            return PyNULL
+        end
+        np = pyimport("numpy")
+        # simple scalar type
+        if pyisnull(descr)
+            return np.dtype(typestr)
+        end
+        # We could juse use np.dtype(descr), but when there is padding, np.dtype(descr)
+        # changes the names of the padding fields from "" to "f{N}". Using this other
+        # dtype constructor avoids this issue and preserves the invariant:
+        #   np.dtype(eltype(array)) == np.array(array).dtype
+        names = []
+        formats = []
+        offsets = []
+        for i = 1:fieldcount(self)
+            nm = fieldname(self, i)
+            push!(names, nm isa Integer ? "f$(nm-1)" : String(nm))
+            ts, ds = pytypestrdescr(fieldtype(self, i))
+            push!(formats, pyisnull(ds) ? ts : ds)
+            push!(offsets, fieldoffset(self, i))
+        end
+        return np.dtype(
+            pydict(
+                names = pylist(names),
+                formats = pylist(formats),
+                offsets = pylist(offsets),
+                itemsize = sizeof(self),
+            ),
+        )
+    end
+    if pyisnull(ans)
+        errset(pybuiltins.AttributeError, "__numpy_dtype__")
+    end
+    return ans
+end
+
 function init_any()
     jl = pyjuliacallmodule
     pybuiltins.exec(
@@ -573,6 +616,9 @@ class Jl(JlBase2):
         return self._jl_callmethod($(pyjl_methodnum(pyjlany_call_nogil)), args, kwargs)
     def _repr_mimebundle_(self, include=None, exclude=None):
         return self._jl_callmethod($(pyjl_methodnum(pyjlany_mimebundle)), include, exclude)
+    @property
+    def __numpy_dtype__(self):
+        return self._jl_callmethod($(pyjl_methodnum(pyjltype_numpy_dtype)))
 """,
             @__FILE__(),
             "exec",